requirements.txt
# Core dependencies for TARS
python-dotenv==1.2.1
Flask==3.1.2

# Twilio for phone calls
twilio==9.9.0

# Gemini AI with native audio support
google-genai>=0.2.0

# Claude AI for programming tasks
anthropic>=0.40.0

# WebSocket support for Twilio Media Streams
websockets==14.1

# Audio processing (Python 3.13+ compatibility)
# audioop was removed in Python 3.13, use audioop-lts as replacement
audioop-lts>=0.2.1

# Async support
aiohttp==3.13.2

# Background task queue for autonomous programming
rq>=1.15.0
redis>=5.0.0

# Logging and utilities
requests==2.32.5
certifi==2025.11.12

# GitHub integration for programmer agent
PyGithub==2.1.1
gitpython==3.1.40

# Headless browser for web automation (WebBrowserAgent, DeepResearchAgent)
playwright>=1.40.0

# Chat functionality
python-telegram-bot>=21.0
discord.py>=2.3.2
